pyproject.toml
[project]
name = "yuketang-scraper"
version = "0.1.0"
description = "Scrape conversation data from yuketang.cn AI workspace and convert to OpenAI format"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.14.0",
    "playwright>=1.40.0",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "tenacity>=9.1.2",
]
